build(client-test-utils): set mocha as devDep test-utils does not have a runtime dependency on mocha. Only types are imported for mocha and those were already set as a devDependency.
Add notes about the test timeout conditional side-effect of loading that is present in test-utils package.

Additionally refactor event and error logger away from testObjectProvider.ts to further isolate modules that depend on the side-effect related code. Nothing strictly relies on the side-effect, but many components reference the module with the side-effect. Per the barrel pattern, any package load will enact the side-effect at this time (assuming no tree shaking).

test-utils is not uniformly used throughout the client workspace which makes it non-trivial to relocate the test timeout logic to @fluid-internal/mocha-test-setup that is broadly used.
