+++ packages/test/test-utils/package.json
@@ -134,7 +134,6 @@
 		"@fluidframework/telemetry-utils": "workspace:~",
 		"best-random": "^1.0.0",
 		"debug": "^4.3.4",
-		"mocha": "^11.7.5",
 		"uuid": "^11.1.0"
 	},
 	"devDependencies": {
@@ -158,6 +157,7 @@
 		"diff": "^3.5.0",
 		"eslint": "~9.39.1",
 		"jiti": "^2.6.1",
+		"mocha": "^11.7.5",
 		"mocha-multi-reporters": "^1.5.1",
 		"rimraf": "^6.1.3",
 		"typescript": "~5.4.5"
