+++ packages/test/test-utils/src/index.ts
@@ -3,8 +3,12 @@
  * Licensed under the MIT License.
  */
 
+export type { IEventAndErrorTrackingLogger } from "./eventAndErrorLogger.js";
+export {
+	EventAndErrorTrackingLogger,
+	getUnexpectedLogErrorException,
+} from "./eventAndErrorLogger.js";
 export { IProvideTestFluidObject, ITestFluidObject } from "./interfaces.js";
-export { LoaderContainerTracker } from "./loaderContainerTracker.js";
 export {
 	fluidEntryPoint,
 	LocalCodeLoader,
@@ -30,25 +34,30 @@ export {
 	TestFluidObjectFactory,
 	TestDataObjectKind,
 } from "./testFluidObject.js";
-export {
-	createDocumentId,
-	DataObjectFactoryType,
-	EventAndErrorTrackingLogger,
-	type IEventAndErrorTrackingLogger,
-	getUnexpectedLogErrorException,
+
+// #region Exports with load side-effect
+// These exports transitively or directly load timeoutUtils.ts that has load
+// side effect of patching Mocha's timeout handling. For the side effect to
+// load, consumer must also load mocha-test-setup, which is pervasively used.
+export { LoaderContainerTracker } from "./loaderContainerTracker.js";
+export type {
 	IDocumentIdStrategy,
 	IOpProcessingController,
 	ITestContainerConfig,
 	ITestObjectProvider,
+} from "./testObjectProvider.js";
+export {
+	createDocumentId,
+	DataObjectFactoryType,
 	TestObjectProvider,
 	TestObjectProviderWithVersionedLoad,
 } from "./testObjectProvider.js";
+export type { SummaryInfo } from "./TestSummaryUtils.js";
 export {
 	createSummarizer,
 	createSummarizerCore,
 	createSummarizerFromFactory,
 	summarizeNow,
-	SummaryInfo,
 } from "./TestSummaryUtils.js";
 export {
 	timeoutAwait,
@@ -63,6 +72,8 @@ export {
 	getContainerEntryPointBackCompat,
 	getDataStoreEntryPointBackCompat,
 } from "./containerUtils.js";
+// #endregion
+
 export {
 	type ContainerRuntimeFactoryWithDefaultDataStoreConstructor,
 	type ContainerRuntimeFactoryWithDefaultDataStoreProps,
